Move refreshAuthentication to dedicated module.
